+++ src/drawable.rs
@@ -35,6 +35,7 @@ pub fn stroke(color: Color) -> DrawType {
 #[derive(Debug, Eq, PartialEq)]
 pub struct Drawable<T: Clone> {
     obj: T,
+    current_rotation: isize,
     draw_type: DrawType,
     drawing_points: Vec<Coord>,
 }
@@ -52,6 +53,10 @@ impl<T: Clone> Drawable<T> {
     pub fn drawing_points(&self) -> &Vec<Coord> {
         &self.drawing_points
     }
+    #[inline]
+    pub fn rotation(&self) -> isize {
+        self.current_rotation
+    }
 }
 
 impl<T: Clone> Drawable<T> {
@@ -61,6 +66,7 @@ impl<T: Clone> Drawable<T> {
             obj,
             draw_type,
             drawing_points,
+            current_rotation: 0,
         }
     }
 }
@@ -72,9 +78,9 @@ impl<T: Clone> Drawable<T> {
 }
 
 impl<T> Drawable<T>
-where
-    Self: CreateDrawable<T>,
-    T: Shape + Clone,
+    where
+        Self: CreateDrawable<T>,
+        T: Shape + Clone,
 {
     pub fn with_translation<P: Into<Coord>>(&self, delta: P) -> Drawable<T> {
         let moved = self.obj.translate_by(delta);
@@ -86,8 +92,15 @@ where
         Drawable::from_obj(moved, self.draw_type)
     }
 
-    pub fn with_rotation<P: Into<Coord>>(&self, degrees: isize) -> Drawable<T> {
+    pub fn with_rotation(&self, degrees: isize) -> Drawable<T> {
         let rotated = self.obj.rotate(degrees);
+        let mut drawable = Drawable::from_obj(rotated, self.draw_type);
+        drawable.current_rotation += degrees;
+        drawable
+    }
+
+    pub fn clear_rotation(&self) -> Drawable<T> {
+        let rotated = self.obj.rotate(-self.current_rotation);
         Drawable::from_obj(rotated, self.draw_type)
     }
 }